Require auth for admin and add user passwords
