+++ src/app/api/users/route.ts
@@ -3,7 +3,8 @@ import { prisma } from '@/lib/prisma'
 
 export async function GET() {
   const users = await prisma.user.findMany({
-    select: { id: true, name: true, email: true, role: true, modules: true }
+    where: { role: { in: ['admin', 'staff'] } },
+    select: { id: true, name: true, email: true, role: true, modules: true },
   })
   return NextResponse.json({ users })
 }
